% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{alphaTriangles}
\alias{alphaTriangles}
\title{Calculate Delauney triangles for alpha hull.}
\usage{
alphaTriangles(x, y, crs)
}
\description{
Performs Delauney triangulation on the provided points
and returns the triangles as a simple features collection.
}
