% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoo.R
\name{aooBuf}
\alias{aooBuf}
\title{Area of Occupancy (AOO) point buffer method}
\usage{
aooBuf(thepoints, bufferradius = 1128, returnV = "S")
}
\arguments{
\item{thepoints}{dataframe of points of x,y}

\item{bufferradius}{in metres, (default = 1128 m, therefore circular area = 2 km2)}

\item{returnV}{three switches for return values 
S = returns the area of the buffered output
SF = return the simple feature for plotting and mapping}
}
\value{
AOO area in km2
}
\description{
Calculates the AOO area using the point buffer method. After Breiner, F. T., & Bergamini, A. (2018), they suggest this avoids issues with grid origin.
}
\examples{
#Build some normally distributed point data around the Troodos mountains ~ 10 km diameter
thepoints <- ptsNormal(50,0.1)
#shift to Troodos mountaions
thepoints <- data.frame(long = thepoints$X + 32.8794, lat = thepoints$Y + 34.9220)
#project the points
ppts <- simProjWiz(thepoints)
#get the number of sub-pop/locations using default method
aooBuf(ppts)
#get the sf object and plot
aoobufs <- aooBuf(ppts,returnV="SF")
plot(aoobufs)
#user defined buffer distance in this case 2 km
aooBuf(ppts,bufferradius=2000,returnV="S")
}
\references{
Breiner, F. T., & Bergamini, A. (2018). Improving the estimation of area of occupancy for IUCN Red List assessments by using a circular buffer approach. Biodiversity and Conservation, 27(9), 2443-2448.
}
\seealso{
\code{\link{subLocGrid}} which uses the same method, but different default radius and also calculates the number of sub-populations or locations
}
\author{
Justin Moat. J.Moat@kew.org
}
