% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoo.R
\name{aooFixedGrid}
\alias{aooFixedGrid}
\title{Optimal shifting grid, Area of Occupancy (AOO)}
\usage{
aooFixedGrid(thepoints, cellsize = 2000, returnV = "S")
}
\arguments{
\item{thepoints}{dataframe of points in metres i.e. c(x,y)}

\item{cellsize}{width of cell in metres (default 2000 m)}

\item{returnV, }{switches to return different sets of results:
S = Simple, returns just the minimum are in km2, (DEFAULT)  
E = Expended simple, returns the solution for the smallest AOO  as list of (area,number of cells, rotation (0 in this case), shift in x direction, shift in y direction) 
ALL = returns a dataframe of all of the results with (number of cells, rotation (0 in this case), shift in x direction (metres), shift in y direction (metres))  
SF = returns a polygon simple feature for mapping, plotting in ggplot or export to GIS systems}
}
\value{
dependent on switch, default is area in km2
}
\description{
Calculates the optimal (smallest) Area of Occupancy AOO  by shifting the grid in x and y direction only.
The minimum solution will be achieved but large point datasets (i.e. over 70 points) will take some time to process.
Processing time is proportional to nop^2 (number of points squared).
Please cite below if using this algorithm:  
Moat, J., Bachman, S. P., Field, R., & Boyd, D. S. (2018). Refining area of occupancy to address the modifiable areal unit problem in ecology and conservation. Conservation biology, 32(6), 1278-1289.
}
\note{
will give a warning if large dataset are used, if this take too long think about using: aooFixedRotation
}
\examples{
#Build and project some points
thepoints <- squareOfPs(19,0.1)
names(thepoints) <- c("lat","long")
thepoints <- simProjWiz(thepoints)
attr(thepoints,'crs')
cellsize = 2000

#just get area in km2
aooFixedGrid(thepoints,cellsize)
#extended return
aooFixedGrid(thepoints,cellsize,"E")
#Build polygons for plotting
library(ggplot2)
polys <- aooFixedGrid(thepoints,cellsize,"SF")
ggplot(data=polys) + geom_sf(color="black",fill="red") +geom_point(data=thepoints,aes(x,y))
#get it all results for all iterations
resultsdf <- aooFixedGrid(thepoints,cellsize,"ALL")
#NB as this is the optimal algorithm, do not expect the histogram to be normal
ggplot(data=resultsdf,aes(nofcells)) + geom_histogram()
#range of cellsize returned
range(resultsdf$nofcells)
#Example of building polygons from one set of results
#plotting the result with the highest AOO
worstgrid <- resultsdf[which.max(resultsdf$nofcells),]
worstpoly <- buildCellPolys_rxy(thepoints,cellsize,worstgrid$rotation,worstgrid$xshift,worstgrid$yshift)
ggplot(data=worstpoly) + geom_sf(color="black",fill="red") +geom_point(data=thepoints,aes(x,y))
}
\references{
Moat, J., Bachman, S. P., Field, R., & Boyd, D. S. (2018). Refining area of occupancy to address the modifiable areal unit problem in ecology and conservation. Conservation biology, 32(6), 1278-1289.

Bachman, S., Moat, J., Hill, A.W., de Torre, J., Scott, B., 2011. Supporting Red List threat assessments with GeoCAT: geospatial conservation assessment tool. Zookeys 126, 117–26. doi:10.3897/zookeys.150.2109
}
\seealso{
\code{\link{aooRating}} for AOO Ratings

\code{\link{aoo}} for simple AOO method

\code{\link{aooFixedRotation}} for systematic methods

\code{\link{buildCellPolys_rxy}} for building grid polygons from points, rotation and shift
}
\author{
Justin Moat. J.Moat@kew.org
}
