% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateDistances}
\alias{calculateDistances}
\title{Calculate the distance between all vertices of a triangle}
\usage{
calculateDistances(triangle)
}
\description{
This is a utility to calculate the edge lengths of a triangle
in a Delauney triangulation.
}
