% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{countrylist}
\alias{countrylist}
\title{PRE-ALPHA Country list from a set of points in geographic coordinates}
\usage{
countrylist(thepointsll)
}
\arguments{
\item{thepoints}{dataframe of points ie c(x,y) in lat long}
}
\value{
list of ISO3 codes
}
\description{
Calculates the country Occurances from a set of points in lat long
}
\details{
calculates the MER of a set of numbers'
}
\examples{
#built set of points
thepoints <- ptsNormal(29,0.1)
#shift to Peru, Colombia and Brazil tripoint
thepoints <- data.frame(long = thepoints$X - 69.9465972, lat = thepoints$Y - 4.2260167)
countrylist(thepoints)
}
\author{
Justin Moat. J.Moat@kew.org
}
