% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{doughnutOfPs}
\alias{doughnutOfPs}
\title{Set of annulus (doughnut) points}
\usage{
doughnutOfPs(noP = 100, gsize = 1, holes = 0.4, aspectRatio = 1, rot = 0)
}
\arguments{
\item{gsize}{scale of area (width of longest size)}

\item{holes}{hole size 0 = none, 1 = ring of points (default=0.4)}

\item{aspectRatio}{i.e. (major axis)/(minor axis), greater than 1, but if <1 it will just switch the axis. For a circle = 1 (default=1)}

\item{rot}{angle of rotation in radians (default = 0)}

\item{nop}{number of points}
}
\value{
dataframe of points (x,y)
}
\description{
Builds random a set of random points for an annulus (doughnut)
}
\examples{
doughnutofpoints <- doughnutOfPs (100,1,0.4,0.5,0.5)
plot(doughnutofpoints)
}
\author{
Justin Moat. J.Moat@kew.org
}
