% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ne.R
\name{get_ne}
\alias{get_ne}
\title{Get Natural Earth country data}
\usage{
get_ne()
}
\value{
sf object with data on global countries
}
\description{
Function to get the Natural Earth country boundary spatial data
}
\keyword{Earth}
\keyword{Natural}
