% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tdwgl3.R
\name{get_tdwgl3}
\alias{get_tdwgl3}
\title{Get TDWG World Geographical Scheme for Recording Plant Distributions (WGSRPD) data}
\usage{
get_tdwgl3()
}
\value{
geojson object with TDWG level 3 boundaries
}
\description{
Function to get the WGSRPD (TDWG) spatial data
}
\keyword{TDWG}
\keyword{WGSRPD}
