% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projWiz.R
\name{llCheck}
\alias{llCheck}
\title{Check point data and latitude and longitude are sensible to work with}
\usage{
llCheck(thepoints)
}
\arguments{
\item{thepoints}{set of points as a dataframe with latitude and longitude}
}
\value{
nothing or warning or error
}
\description{
Checked the dataframe for NA, latitude below -90 or above 90, longitude below -180 and above 180, also warns if it finds whole numbers or 0
}
\note{
this used as a quick check to see if data can be worked with and are on the earth. 
If coordinate cleaning is required we suggest you review https://cran.r-project.org/web/packages/CoordinateCleaner/index.html
}
\examples{
# build dataframe
lat <- runif (200,-24,-12)
long <- runif (200,43,51)
llp <- data.frame(lat,long)
#add some odd data
llp[1:10,] <- c(10,10)
llp[11:100,] <- c(0,0)
#below will return two warnings
llcheck(llp)
llp[101,]<- NA
#below will return an error
llcheck(llp)
llp[101,]<- c(1000,-20)
#below will return an error
llcheck(llp)
llp[101,]<- c(-20,1000)
#below will return an error
llcheck(llp)
}
\author{
Justin Moat. J.Moat@kew.org
}
