% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{longestAxis}
\alias{longestAxis}
\title{Longest distance from a set of points}
\usage{
longestAxis(thepoints, returnV = "S")
}
\arguments{
\item{thepoints}{dataframe of points of x,y}

\item{returnV, }{two switches either S for simply the distance or P for a dataframe of the two furthest points}
}
\value{
distance in metres or two points for the longest distance
}
\description{
Calculates the longest distances from a set of points
}
\note{
Useful as a scale for cellsize and location buffers, Willis et al 2003 suggest 1/10 of this for cellsize for AOO calculations as does Rivers et al (2010) for buffer distance for sub-population or location calculations.
}
\examples{
x <- runif (20,0,10)
y <- runif (20,0,10)
df <- data.frame(x,y) 
#distance only
longestAxis(df)
#two furthest points
dp <- longestAxis(df,'P')
plot(df, asp=1)
points(dp,pch=16)
}
\references{
Willis, F., Moat, J., & Paton, A. (2003). Defining a role for herbarium data in Red List assessments: a case study of Plectranthus from eastern and southern tropical Africa. Biodiversity & Conservation, 12(7), 1537-1552.
}
\author{
Justin Moat. J.Moat@kew.org
}
