% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MER.R
\name{MER}
\alias{MER}
\title{Minimum Enclosing Rectangle (MER)}
\usage{
MER(thepoints)
}
\arguments{
\item{thepoints}{dataframe of points ie c(x,y)}
}
\value{
list of 4 doubles = xmin,xmax,ymin,ymax
}
\description{
Calculates the minimum enclosing rectangle (MER) from a set of points (x,y)
}
\details{
calculates the MER of a set of numbers'
}
\examples{
x <- runif (20,0,10)
y <- runif (20,0,10)
df <- data.frame(x,y) 
MER (df)
}
\author{
Justin Moat. J.Moat@kew.org
}
