% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{Mode}
\alias{Mode}
\title{Mode of a test of points}
\usage{
Mode(x)
}
\arguments{
\item{x}{set of numbers}
}
\value{
number
}
\description{
Calculates the mode (the value that occurs most often) of a set of data - note the capital M
}
\details{
calculates the mode of a set of numbers'
}
\note{
Orginally from http://stackoverflow.com/questions/2547402/standard-library-function-in-r-for-finding-the-mode
}
\examples{
a <- c(5,5,5,6,7,8,9)
Mode(a)
}
\author{
Justin Moat. J.Moat@kew.org
}
