% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalofPs}
\alias{normalofPs}
\title{Set of normally distributed points}
\usage{
normalofPs(nop = 100, gsize = 1)
}
\arguments{
\item{nop}{number of points}

\item{gsize}{scale of area (width of longest size), NB this is within 3 SD so some point may extend beyond this}
}
\value{
dataframe of points (x,y)
}
\description{
Builds random a set of normally distribute
}
\examples{
normalpoints <- normalofPs(100,1)
plot(normalpoints)
}
\author{
Justin Moat. J.Moat@kew.org
}
