% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ovalOfPs}
\alias{ovalOfPs}
\title{Set of a circle or oval points}
\usage{
ovalOfPs(noP, gsize, rot, aspectRatio = 1)
}
\arguments{
\item{gsize}{size of square (width of longest size)}

\item{rot}{angle of rotation in radians}

\item{aspectRatio}{i.e. (major axis)/(minor axis), greater than 1, but if <1 it will just switch the axis. For a circle = 1}

\item{nop}{number of points}
}
\value{
dataframe of points (x,y)
}
\description{
Builds a random set of points for a circular or oval area
}
\examples{
plot(ovalOfPs(100,1,1,0.5),asp=1)
}
\author{
Justin Moat. J.Moat@kew.org
}
