% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ptsDoughnut}
\alias{ptsDoughnut}
\title{Builds a set of annulus (doughnut) points}
\usage{
ptsDoughnut(noP = 100, gsize = 1, holes = 0.4, aspectRatio = 1, rot = 0)
}
\arguments{
\item{gsize}{scale of area (~ width of longest axis)}

\item{holes}{hole size, between 0 and 1, 0 = no hole, 1 = cirlce of points (default=0.4)}

\item{aspectRatio}{For an ellipse shape i.e. (major axis)/(minor axis), greater than 1, but if <1 it will just switch the axis. For a circle = 1 (default=1)}

\item{rot}{angle of rotation for the ellipse in radians (default = 0)}

\item{nop}{number of points}
}
\value{
dataframe of points (x,y)
}
\description{
Builds random a set of random points for an annulus (doughnut shape)
}
\examples{
doughnutofpoints <- ptsDoughnut (100,1,0.4,0.5,0.5)
plot(doughnutofpoints)
}
\author{
Justin Moat. J.Moat@kew.org
}
