% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ptsOval}
\alias{ptsOval}
\title{Builds a set of a circle or oval points}
\usage{
ptsOval(noP, gsize = 2000, rot = 0.785398, aspectRatio = 1)
}
\arguments{
\item{gsize}{size of square (width of longest size) (default 2000)}

\item{rot}{angle of rotation in radians (default 0.785398)}

\item{aspectRatio}{i.e. (major axis)/(minor axis), can be greater than 1, but if so it will just switch the axis. (default =1 ie circle).}

\item{nop}{number of points}
}
\value{
dataframe of points (x,y)
}
\description{
Builds a random set of points for a circular or oval area
}
\examples{
plot(ptsOval(100,1,1,0.5),asp=1)
}
\author{
Justin Moat. J.Moat@kew.org
}
