% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IUCNcat.R
\name{ratingPop}
\alias{ratingPop}
\title{IUCN rating based on population reduction}
\usage{
ratingPop(pReduction, subCr = 2, abb = TRUE)
}
\arguments{
\item{pReduction}{reduction as a percentage}

\item{subCr}{sub Criteria category: 1 or 2 or 3 or 4 (2 default)}

\item{abb}{abbreviation TRUE or FALSE , TRUE = 2 letter code, FALSE = full text (see value), default = TRUE}
}
\value{
Text
one of CR, EN, VU, NT, LC or Critically Endangered, Endangered, Vulnerable, Near Threatened, Least Concern
}
\description{
Calculates IUCN rating based on based on population reduction as a percentage
}
\details{
Calculates IUCN rating based on population reduction
}
\note{
subCr 1 use the limits 90%,70%,50%,25% for CR,EN,VU,NT

subCr 2,3,4, use the limits 80%,50%,30%,10% for CR,EN,VU,NT

NT category is a guideline
}
\examples{
ratingPop(65,2,TRUE)
ratingPop(65,2,FALSE)
}
\references{
IUCN, 2012. IUCN RED LIST CATEGORIES AND CRITERIA, 2nd ed. IUCN, Gland, Switzerland. 

IUCN Standards and Petitions Subcommittee, 2014. Guidelines for Using the IUCN Red List Categories and Criteria. Version 11.

IUCN Standards and Petitions Subcommittee, 2016. Guidelines for Using the IUCN Red List Categories and Criteria. Version 12.
}
\author{
Justin Moat. J.Moat@kew.org
}
