% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{squareOfPs}
\alias{squareOfPs}
\title{Set of points in a square area}
\usage{
squareOfPs(noP, gsize)
}
\arguments{
\item{gsize}{size of square (width or height)}

\item{nop}{number of points}
}
\value{
dataframe of points (x,y)
}
\description{
Builds a random set of points for a square area
}
\examples{
dfofpoints <- squareOfPs(100,1)
}
\author{
Justin Moat. J.Moat@kew.org
}
