% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_loc_frag.R
\name{subLocAlpha}
\alias{subLocAlpha}
\title{Sub-populations or locations using the alpha hull method}
\usage{
subLocAlpha(
  thepoints,
  barrierDis = longestAxis(thepoints, returnV = "S")/10,
  returnV = "S"
)
}
\arguments{
\item{thepoints}{dataframe of points of x,y}

\item{barrierDis}{in metres, (default = 1/10 the longest axis of the set of points)}

\item{returnV}{four switches either \cr
S = for simply the number of sub-pop or locations \cr
AREA = returns the area of the cells (use AOO or EOO for useful areas for IUCN assessment) \cr
SF = returns a multipolygon simple feature of the alpha hull, for mapping, plotting in ggplot or export to a GIS  \cr
ALL = returns a multipolygon of all the triangular elements of the alpha hull}
}
\value{
number of sub-pop/locations
}
\description{
Calculates the number of Sub-population or  locations using the alpha hull method
}
\note{
Malin et al 2010 suggest barrier width of 1/10 the longest distance.

As the barrier distance increases for Alpha hulls, point become outliners with no area. This algorithm returns the number of grouped points (in the hull) and the number of these outlying points for the sub population or location calculations. 
We would not recommend using Alpha hulls for sub populations or locations, but it is supplied here for consistency and for those that want to experiment.
}
\examples{
#Build some normally distributed point data around the Troodos mountains ~ 10 km diameter
thepoints <- ptsNormal(50,0.1)
#shift to Troodos mountaions
thepoints <- data.frame(long = thepoints$X + 32.8794, lat = thepoints$Y + 34.9220)
#project the points
ppts <- simProjWiz(thepoints)
#get number of subpop with defaults
subLocAlpha(ppts)
#get area
subLocAlpha(ppts,returnV = "AREA")
#user defined distance
subLocAlpha(ppts,barrierDis=2000)
subpop <- subLocAlpha(ppts,returnV = "SF")
#plotting the results
library(ggplot2)
ggplot(data=subpop) + geom_sf() + geom_point(data=ppts,aes(X,Y))

}
\references{
Rivers, M. C., Bachman, S. P., Meagher, T. R., Lughadha, E. N., & Brummitt, N. A. (2010). Subpopulations, locations and fragmentation: applying IUCN red list criteria to herbarium specimen data. Biodiversity and Conservation, 19(7), 2071-2085. https://link.springer.com/article/10.1007/s10531-010-9826-9
}
\seealso{
\code{\link{longestAxis}} to calculate the length of the longest axis

\code{\link{aHullMean}} to calculate the alpha hull using multiples of the mean lengths of the Delaunay triangulation
}
\author{
Justin Moat. J.Moat@kew.org
}
