% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sub_loc_frag.R
\name{subLocGrid}
\alias{subLocGrid}
\title{Sub-population or number of locations using grid adjacency method}
\usage{
subLocGrid(
  thepoints,
  cellwidth = longestAxis(thepoints, returnV = "S")/10,
  neighborhood = "queen",
  returnV = "S"
)
}
\arguments{
\item{thepoints}{dataframe of points of x,y}

\item{cellwidth}{in metres, (default = 1/10 the longest axis of the set of points)}

\item{neighborhood}{either rook (up,down,left and right) or queen (rook plus diagonals) default = queen}

\item{returnV}{three switches for return values \cr
S = simply the number of sub-pop or locations \cr
AREA = returns the area of the cells (use AOO for useful areas methods for IUCN assessments) \cr
SF = return the simple feature for plotting, mapping and export to a GIS
#' @return number of sub-pop/locations}
}
\description{
Calculates the number of Sub-population or number of locations using the grid adjacency method
}
\note{
Malin et al 2010 suggest cells width of 1/10 the longest distance. This uses the default grid (0,0)
}
\examples{
#Build some normally distributed point data around the Troodos mountains ~ 10 km diameter
thepoints <- ptsNormal(50,0.1)
#shift to Troodos mountaions
thepoints <- data.frame(long = thepoints$X + 32.8794, lat = thepoints$Y + 34.9220)
#project the points
ppts <- simProjWiz(thepoints)
#get the number of sub-pop/locations using default method
subLocGrid(ppts)
#get area
subLocGrid(ppts,returnV="AREA")
#get the SF object and plot it
p <- subLocGrid(ppts,returnV="SF")
plot (p)
Using rook neighborhood
subLocGrid(ppts,neighborhood = "rook")
#with user defined cellwidth
subLocGrid(ppts,neighborhood = "rook",cellwidth = 2000)
}
\references{
Rivers, M. C., Bachman, S. P., Meagher, T. R., Lughadha, E. N., & Brummitt, N. A. (2010). Subpopulations, locations and fragmentation: applying IUCN red list criteria to herbarium specimen data. Biodiversity and Conservation, 19(7), 2071-2085. https://link.springer.com/article/10.1007/s10531-010-9826-9
}
\seealso{
\code{\link{longestAxis}} to calculate the length of the longest axis

\code{\link{aoo}} to calculate simple AOO or

\code{\link{aooFixedGrid}} to calculate optimal AOO on a non-rotating grid or

\code{\link{aooFixedRotation}}to calculate optimal AOO with rotation and shift
}
\author{
Justin Moat. J.Moat@kew.org
}
