% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RapoportMP.R
\name{subLocRapoport}
\alias{subLocRapoport}
\title{Sub-population or number of locations using Rapoport's mean propinquity}
\usage{
subLocRapoport(thepoints, barrierDis, bufferDis, returnV = "S")
}
\arguments{
\item{thepoints}{dataframe of points in metres i.e. c(X,Y)}

\item{barrierDis}{distance (in m) to be used to define barriers (if not defined the default is 2 x mean branch length of the Euclidean Minimum spanning tree). Rivers et al (2010) and Willis et al (2003) suggest 1/10 of the longest Axis may be useful.}

\item{bufferDis}{distance (in m) to be used to buffer points and connected edges (if not defined the default is the mean branch length of the Euclidean Minimum spanning tree )}

\item{returnV, }{switches to return different sets of results: \cr
S = Simple, returns just the number of sub-populations or locations (default) \cr
AREA = returns the area (km2) of the buffered points and tree \cr
SF = returns a list with two simple features (tree and buffers) for visualisation, mapping, further analysis and export to a GIS}
}
\value{
dependent on switch, default is area in km2

see returnV
}
\description{
Calculates Rapoport's (1982) mean propinquity from a set of points. 
For details on Rapoport’s mean propinquity methodsee Willis et al. 2003 and Rapoport 1982. This technique is
based on a Euclidean minimum spanning tree (eMST), which is a
set of lines (branches) that connects all points in the minimum possible distance. Sub-populations
are defined when the branches (line from a pair of points) distance is greater than the defined distance 
(Rapoport suggests twice the mean edge distance (Willis et al. 2003)). 
Area is defined by  buffering this tree by a defined distance (Rapoport suggests the mean) 
any isolated populations (single populations or points) are also buffered by this distance.
}
\examples{
#Build and project some points
thepoints <- ptsSquare(19,0.1)
names(thepoints) <- c("lat","long")
thepoints <- simProjWiz(thepoints)
#number of sub-populations/locations using default distances
subLocRapoport(thepoints)
#defining your own distances
subLocRapoport(thepoints,barrierDis = 2000,bufferDis = 1000)
#area of sub populations/locations using default distances
subLocRapoport(thepoints,returnV = "AREA")
#results as simple features for plotting
sfs <- subLocRapoport(thepoints,returnV = "SF")
library(ggplot2)
ggplot (data=sfs$tree) + geom_sf(aes(color=barrier)) + geom_sf(data=sfs$buffers,colour='green', fill=NA) + geom_point(data=thepoints,aes(X,Y)) 
#get the mean of distance for the tree branches
mean(sfs$tree$distance)

}
\references{
https://github.com/allanvc/emstreeR

March, W. B., Ram, P., & Gray, A. G. (2010, July). Fast euclidean minimum spanning tree: algorithm, analysis, and applications. In Proceedings of the 16th ACM SIGKDD international conference on Knowledge discovery and data mining (pp. 603-612).

Rivers, M. C., Bachman, S. P., Meagher, T. R., Lughadha, E. N., & Brummitt, N. A. (2010). Subpopulations, locations and fragmentation: applying IUCN red list criteria to herbarium specimen data. Biodiversity and Conservation, 19(7), 2071-2085. https://link.springer.com/article/10.1007/s10531-010-9826-9

Willis, F., Moat, J., & Paton, A. (2003). Defining a role for herbarium data in Red List assessments: a case study of Plectranthus from eastern and southern tropical Africa. Biodiversity & Conservation, 12(7), 1537-1552.

Moat J (2007) Conservation assessment tools extension for ArcView 3.x, version 1.2. GIS Unit, Royal Botanic Gardens, Kew, UK. https://www.kew.org/sites/default/files/2019-02/Conservation_assessment_tools_extension.pdf

Rapoport E.H. 1982. Areography: Geographical Strategies of Species. Pergamon Press, New York.
}
\seealso{
\code{\link{eMST}} Euclidean Minimum spanning tree

\code{\link{longestAxis}} Longest distance from a set of points
}
\author{
Justin Moat. J.Moat@kew.org
}
