% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdwglist.R
\name{tdwglist}
\alias{tdwglist}
\title{Extract wgsrpd regions from a set of points in lat long}
\usage{
tdwglist(thepointsll, returnV = "S")
}
\arguments{
\item{thepoints}{dataframe of points ie c(x,y) in lat long}
}
\value{
dataframe or sf object with wgsrpd regions
}
\description{
Extract wgsrpd regions from a set of points in lat long
}
\examples{
#built set of points
thepoints <- ptsNormal(29,0.1)
#shift to Peru, Colombia and Brazil tripoint
thepoints <- data.frame(long = thepoints$X - 69.9465972, lat = thepoints$Y - 4.2260167)
tdwglist(thepoints)
}
