% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mer}
\alias{mer}
\title{Minimum Enclosing Rectangle (mer)}
\usage{
mer(thepoints)
}
\arguments{
\item{thepoints}{dataframe of points ie c(x,y)}
}
\value{
vector of 4 doubles = xmin,xmax,ymin,ymax
}
\description{
Calculates the minimum enclosing rectangle (mer) from a set of points (x,y)
}
\details{
calculates the MER of a set of numbers'
}
\examples{
x <- runif (20,0,10)
y <- runif (20,0,10)
df <- data.frame(x,y) 
mer (df)
}
\author{
Justin Moat. J.Moat@kew.org
}
